Added a connection Class/Object and key hashing option
Created a separate class to establish connection to redis. Class makes
it a bit easier to create connections where custom settings are
required, such as DB #, host, port. Also added an option to generate a
md5 digest of the pickle object generated as a result of pickling
arguments to a function. This reduces chance of collisions in key
namespace and makes for consistent length of keys.