Minor improvements to RedisConnect object
Removed assumptions about always having a successful connection, and
added a try/except block to handle instances where connection fails.
Modifed decorators to simply return the function instead of trying to
actually cache data if the cache is unavailable. We check
self.connection, and if == None, we just return the function we were
going to wrap.