Add DoNoCache exception
The DoNotCache exception is for use with the cache_it decorator to tell
it to not cache this result, but still allow the result to be returned.
It also wraps the storer call in a try/except catching the
redis.ConnectionError. Since this is a cache decorator it should not
throw an exception on a ConnectionError it should just log it and move
on returning the result